post deletion should have a higher chance of being workable now

+++ app/services/PostService.js
@@ -10,12 +10,13 @@
   };
   AbstractService = require('./AbstractService.js');
   module.exports = function(dependencies) {
-    var PostService, Thread, Tracker, ValidationError, hashlib, images, janitor, mongoose, promise, succeed, _ref;
+    var PostService, PreconditionError, Thread, Tracker, ValidationError, hashlib, images, janitor, mongoose, promise, succeed, _ref;
     mongoose = dependencies.mongoose, hashlib = dependencies.hashlib;
     _ref = dependencies.lib('promises'), promise = _ref.promise, succeed = _ref.succeed;
     images = dependencies.lib('images');
     janitor = dependencies.lib('janitor');
     ValidationError = dependencies.lib('errors/ValidationError');
+    PreconditionError = dependencies.lib('errors/PreconditionError');
     Thread = mongoose.model('Thread');
     Tracker = mongoose.model('Tracker');
     return PostService = (function() {
@@ -25,18 +26,24 @@
       }
       PostService.prototype.remove = function(board, id, password) {
         return promise(function(success, error) {
-          return Tracker.find({
+          if (String(password != null ? password : '').length === 0) {
+            return error(new PreconditionError("no password given"));
+          }
+          return Tracker.findOne({
             board: board,
             post: id
           }).run(function(err, tracker) {
             if (err) {
               return error(err);
             }
-            return Thread.find({
+            if (!tracker) {
+              return error(new PreconditionError("no such post"));
+            }
+            return Thread.findOne({
               markedForDeletion: false,
               id: tracker.thread
             }).run(function(err, thread) {
-              var post;
+              var post, _ref2;
               if (err) {
                 return error(err);
               }
@@ -52,16 +59,20 @@
                 }
                 return _results;
               })())[0];
+              if (!post) {
+                return error(new PreconditionError("no such post"));
+              }
               if (!post.password === hashlib.sha1(password)) {
                 return error(new ValidationError("unable to delete post; password does not match"));
               }
+              return error(new Error);
               if (thread.id === post.id) {
                 return janitor.sweepThread(thread).then(succeed(post));
               } else {
                 images.deleteByPost(thread.board, post);
                 tracker.remove();
                 post.remove();
-                if (thread.lastPost.id === postid) {
+                if (((_ref2 = thread.lastPost) != null ? _ref2.id : void 0) === post.id) {
                   thread.lastPost.remove();
                 }
                 thread.save();
